voice_agent/requirements.txt
fastapi
uvicorn
python-multipart
supabase>=2.5.0
python-dotenv
ollama
faster-whisper
kokoro-onnx
soundfile
numpy
requests
